Fix errors in space handling; add manual cleanup file
There were some typos in the script that generated
the cross-references between exercises and errata
in the solutions. We also shortened the text for
these cross-references.

We make an attempt to regularize tho amount
of white space before the errata entries. Sometimes the
original files have three spaces; sometimes they have four.